Fix for #14.
It is not expected to update the dictionary every step. Since python does not remove a dictionary while the code is looping over its items, it cannot raise a runtime error or give a wrong answer, but it is still incorrent.
The same code in C++ would certainly caused SEGFAULT or anything.